# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fursa Platform API - Digital Employment Platform for Gaza
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [me]
      summary: Current authenticated user profile (auto-creates on first call)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentUser
      tags: [me]
      summary: Update profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
  /me/role:
    post:
      operationId: setUserRole
      tags: [me]
      summary: Set role during onboarding (seeker | employer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRoleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
...
  /employer/jobs:
    get:
      operationId: listEmployerJobs
      tags: [employer]
      summary: List jobs posted by current employer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployerJob"
    post:
      operationId: createEmployerJob
      tags: [employer]
      summary: Create a new job (status starts as 'pending')
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJob"
  /employer/jobs/{id}:
    get:
      operationId: getEmployerJob
      tags: [employer]
      summary: Get an employer's own job (any status)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJob"
    patch:
      operationId: updateEmployerJob
      tags: [employer]
      summary: Update a job (resets to pending if approved fields changed)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJob"
    delete:
      operationId: deleteEmployerJob
      tags: [employer]
      summary: Delete a job
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employer/jobs/{id}/toggle-open:
    post:
      operationId: toggleJobOpen
      tags: [employer]
      summary: Open/close a job for new applications
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJob"
  /employer/jobs/{id}/applications:
    get:
      operationId: listJobApplications
      tags: [employer]
      summary: List applications for a job owned by current employer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployerApplication"
  /employer/applications/{id}:
    patch:
      operationId: updateApplicationStatus
      tags: [employer]
      summary: Accept or reject an application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerApplication"
  /employer/applications/{id}/seen:
    post:
      operationId: markApplicationSeen
      tags: [employer]
      summary: Mark application as seen
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Updated
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Role:
      type: string
      enum: [seeker, employer, admin]
    JobType:
      type: string
      enum: [online, field, hybrid]
    JobStatus:
      type: string
      enum: [pending, approved, rejected]
    ApplicationStatus:
      type: string
      enum: [pending, accepted, rejected]
...
    CurrentUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        cvObjectPath:
          type: ["string", "null"]
        isActive:
          type: boolean
        onboarded:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - onboarded
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        cvObjectPath:
          type: ["string", "null"]
    SetRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [seeker, employer]
      required: [role]
...
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        contactInfo:
          type: string
        deadline:
          type: ["string", "null"]
          format: date-time
      required: [title, description, type, category, contactInfo]
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        contactInfo:
          type: string
        deadline:
          type: ["string", "null"]
          format: date-time
    EmployerJob:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        requirements:
          type: ["string", "null"]
        type:
          $ref: "#/components/schemas/JobType"
        category:
          type: string
        contactInfo:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        rejectionReason:
          type: ["string", "null"]
        isOpen:
          type: boolean
        deadline:
          type: ["string", "null"]
          format: date-time
        applicationsCount:
          type: integer
        unseenApplicationsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - type
        - category
        - contactInfo
        - status
        - isOpen
        - applicationsCount
        - unseenApplicationsCount
        - createdAt
    EmployerApplication:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        applicantPhone:
          type: ["string", "null"]
        applicantLocation:
          type: ["string", "null"]
        applicantBio:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        cvObjectPath:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        seenByEmployer:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - applicantName
        - applicantEmail
        - status
        - seenByEmployer
        - createdAt
    UpdateApplicationBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected]
      required: [status]
...