# lib/api-zod/src/generated/types/employerApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationStatus } from "./applicationStatus";

export interface EmployerApplication {
  id: number;
  jobId: number;
  applicantName: string;
  applicantEmail: string;
  /** @nullable */
  applicantPhone?: string | null;
  /** @nullable */
  applicantLocation?: string | null;
  /** @nullable */
  applicantBio?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  cvObjectPath?: string | null;
  status: ApplicationStatus;
  seenByEmployer: boolean;
  createdAt: Date;
}