# lib/api-zod/src/generated/types/updateApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateApplicationBodyStatus } from "./updateApplicationBodyStatus";

export interface UpdateApplicationBody {
  status: UpdateApplicationBodyStatus;
}