# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fursa Platform API - Digital Employment Platform for Gaza
 * OpenAPI spec version: 0.1.0
 */
import type { JobType } from "./jobType";

export interface UpdateJobBody {
  title?: string;
  description?: string;
  /** @nullable */
  requirements?: string | null;
  type?: JobType;
  category?: string;
  contactInfo?: string;
  /** @nullable */
  deadline?: Date | null;
}